add endpoint POST /api/articles and MVC refactor
