+++ controllers/topics.controllers.js
@@ -1,0 +1,11 @@
+const { fetchTopics } = require("../models/topics.models");
+
+module.exports.getTopics = (req, res, next) => {
+  fetchTopics()
+    .then((topics) => {
+      res.status(200).send({ topics });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
