+++ endpoints.json
@@ -57,6 +57,28 @@
       ]
     }
   },
+  "POST /api/articles": {
+    "description": "add an article to the database",
+    "queries": [],
+    "requestBodyFormat": {
+      "author": "rogersop",
+      "title": "why queenie is a cutie",
+      "body": "Even though she is the worst sometimes she is actually very very cute",
+      "topic": "cats",
+      "OPTIONAL article_img_url": "https://http.cat/images/201.jpg"
+    },
+    "exampleResponse": {
+      "author": "rogersop",
+      "title": "why queenie is a cutie",
+      "body": "Even though she is the worst sometimes she is actually very very cute",
+      "topic": "cats",
+      "article_img_url": "https://http.cat/images/201.jpg",
+      "article_id": 13,
+      "votes": 0,
+      "created_at": "2023-05-30T15:59:13.341Z",
+      "comment_count": 0
+    }
+  },
   "GET /api/articles/:article_id/comments": {
     "description": "serves a list of comments by article_id",
     "queries": [],
@@ -151,6 +173,23 @@
         "username": "butter_bridge",
         "name": "jonny",
         "avatar_url": "https://www.healthytherapies.com/wp-content/uploads/2016/06/Lime3.jpg"
+      }
+    }
+  },
+  "PATCH /api/comments/:comment_id": {
+    "description": "serves a single updated comment, by comment_id",
+    "queries": [],
+    "requestBodyFormat": {
+      "inc_votes": 12
+    },
+    "exampleResponse": {
+      "comment": {
+        "comment_id": 1,
+        "body": "Oh, I've got compassion running out of my nose, pal! I'm the Sultan of Sentiment!",
+        "article_id": 9,
+        "author": "butter_bridge",
+        "votes": 26,
+        "created_at": "2020-04-06T12:17:00.000Z"
       }
     }
   }
