+++ models/topics.models.js
@@ -1,0 +1,7 @@
+const db = require("../db/connection");
+
+module.exports.fetchTopics = () => {
+  return db.query("SELECT * FROM topics").then(({ rows }) => {
+    return rows;
+  });
+};
